# api/notificationservice/notificationservice_swagger.yaml
basePath: /api/notification-service
definitions:
...
  mattermostdto.MattermostNotificationChannelRequest:
    properties:
      channelName:
        type: string
      description:
        type: string
      webhookUrl:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API of the Notification service
  license:
    name: AGPL-3.0-or-later
  title: Notification Service API
  version: "1.0"
paths:
...
  /notification-channel/mattermost:
    get:
      description: List mattermost notification channels by type
      parameters:
      - description: Channel type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/mattermostdto.MattermostNotificationChannelRequest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: List Mattermost Channels
      tags:
      - mattermost-channel
    post:
      consumes:
      - application/json
      description: Create a new mattermost notification channel
      parameters:
      - description: Mattermost channel to add
        in: body
        name: MattermostChannel
        required: true
        schema:
          $ref: '#/definitions/mattermostdto.MattermostNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/mattermostdto.MattermostNotificationChannelRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Create Mattermost Channel
      tags:
      - mattermost-channel
  /notification-channel/mattermost/{id}:
    delete:
      description: Delete a mattermost notification channel
      parameters:
      - description: Mattermost channel ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Delete Mattermost Channel
      tags:
      - mattermost-channel
    put:
      consumes:
      - application/json
      description: Update an existing mattermost notification channel
      parameters:
      - description: Mattermost channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Mattermost channel to update
        in: body
        name: MattermostChannel
        required: true
        schema:
          $ref: '#/definitions/mattermostdto.MattermostNotificationChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mattermostdto.MattermostNotificationChannelRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Update Mattermost Channel
      tags:
      - mattermost-channel
  /notification-channel/mattermost/check:
    post:
      consumes:
      - application/json
      description: Check if a mattermost server is able to send a test message
      parameters:
      - description: Mattermost channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Mattermost server test message sent successfully
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - KeycloakAuth: []
      summary: Check mattermost server
      tags:
      - mattermost-channel
...
securityDefinitions:
  KeycloakAuth:
    authorizationUrl: '{{.KeycloakAuthUrl}}/realms/{{.KeycloakRealm}}/protocol/openid-connect/auth'
    flow: implicit
    type: oauth2
swagger: "2.0"